requirements.txt
cx_Oracle==8.3.0
mysqlclient==2.2.4
psycopg2==2.9.9
pyodbc==5.1.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2

python-dotenv~=1.0.1